fix(ci): python full coverage - ws_* plan artifacts were treated as blocking stale overlaps
